requirements.txt
Django==5.2.6
djangorestframework==3.16.1
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
Pillow==10.0.1
python-decouple==3.8

